requirements.txt
python-keycloak==3.9.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
sqlalchemy==2.0.31
fastapi[all]==0.111.0
strawberry-graphql[all]==0.235.0
uvicorn==0.30.1
